requirements.txt
flask==2.0.1
twilio==7.15.4
supabase==1.0.3
python-dotenv==0.19.2
gunicorn==20.1.0
werkzeug==2.0.1
elevenlabs==0.2.24
requests==2.28.1
flask-wtf==1.0.0
pandas==1.3.5
plotly==5.6.0
whitenoise==6.4.0 